+++ dilithium/src/ml_dsa_87.rs
@@ -1,361 +1,9 @@
-use zeroize::{Zeroize, ZeroizeOnDrop, Zeroizing};
+//! ML-DSA-87 (FIPS 204 category 5) public API.
+//!
+//! Thin frontend over the const-generic signing core, instantiated at the
+//! [`crate::params::ml_dsa_87`] parameter set.
 
-use crate::{
-	errors::{KeyParsingError, KeyParsingError::BadSecretKey, SignatureError},
-	params, SensitiveBytes32,
-};
-use core::fmt;
-
-pub const SECRETKEYBYTES: usize = crate::params::SECRETKEYBYTES;
-pub const PUBLICKEYBYTES: usize = crate::params::PUBLICKEYBYTES;
-pub const SIGNBYTES: usize = crate::params::SIGNBYTES;
-pub const KEYPAIRBYTES: usize = SECRETKEYBYTES + PUBLICKEYBYTES;
-
-/// Maximum message size for signing/verification (64 MiB).
-///
-/// This limit prevents denial-of-service attacks via memory exhaustion from
-/// oversized messages. The limit is generous enough for any legitimate use case.
-pub const MAX_MESSAGE_SIZE: usize = 64 * 1024 * 1024;
-
-pub type Signature = [u8; SIGNBYTES];
-
-/// A pair of private and public keys.
-///
-/// `Clone` is intentionally not derived because the embedded `SecretKey` is sensitive.
-/// To explicitly copy a keypair (e.g. to move it into a closure), serialize and
-/// reconstruct: `Keypair::from_bytes(&keypair.to_bytes())?`. This forces the
-/// duplication of secret material to be visible at every call site.
-pub struct Keypair {
-	pub secret: SecretKey,
-	pub public: PublicKey,
-}
-
-impl Keypair {
-	/// Generate a Keypair instance.
-	///
-	/// # Arguments
-	///
-	/// * 'entropy' - bytes for determining the generation process (must be at least 32 bytes)
-	///
-	/// Note: The entropy is moved here and zeroized after use, along with the derived secret key.
-	pub fn generate(entropy: SensitiveBytes32) -> Keypair {
-		let mut pk = [0u8; PUBLICKEYBYTES];
-		let mut sk = [0u8; SECRETKEYBYTES];
-		crate::sign::keypair(&mut pk, &mut sk, entropy);
-		let keypair = Keypair {
-			// Constructed directly rather than via `SecretKey::from_bytes`:
-			// a freshly generated key is consistent by construction, and the
-			// import-path validation would redo the keygen-scale derivation.
-			secret: SecretKey { bytes: sk },
-			public: PublicKey::from_bytes(&pk).expect("Should never fail"),
-		};
-		sk.zeroize();
-		// entropy is automatically zeroized when it drops (ZeroizeOnDrop)
-		keypair
-	}
-
-	/// Convert a Keypair to a bytes array.
-	///
-	/// Returns a self-wiping buffer containing private and public key bytes.
-	/// The buffer contains the full secret key, so it is `Zeroizing`: it is
-	/// erased when dropped, and callers no longer need (or are able to
-	/// forget) manual cleanup. Deref to `[u8; KEYPAIRBYTES]` for access.
-	pub fn to_bytes(&self) -> Zeroizing<[u8; KEYPAIRBYTES]> {
-		// Copy the secret half directly from the stored key rather than via
-		// `self.secret.to_bytes()`: that would materialize a temporary copy
-		// of the whole secret key on the stack.
-		//
-		// The working buffer is deliberately *not* returned by move: a
-		// moved-from local never runs `Drop`, so if the compiler does not
-		// elide the return move (empirically it does not here), the frame
-		// would keep an unwiped copy. Instead the buffer is copied into a
-		// fresh wrapper for the caller and then wiped by its own `Drop` on
-		// scope exit — only the copy into the return slot is left to the
-		// optimizer, and the release-mode stack probe in
-		// `tests/import_stack_zeroization.rs` verifies it is elided.
-		let mut result = Zeroizing::new([0u8; KEYPAIRBYTES]);
-		result[..SECRETKEYBYTES].copy_from_slice(&self.secret.bytes);
-		result[SECRETKEYBYTES..].copy_from_slice(&self.public.to_bytes());
-		Zeroizing::new(*result)
-	}
-
-	/// Create a Keypair from bytes.
-	///
-	/// # Arguments
-	///
-	/// * 'bytes' - private and public keys bytes
-	///
-	/// Returns a Keypair
-	///
-	/// # Consistency check
-	///
-	/// The public half is re-derived from the secret half and must match the
-	/// supplied public-key bytes exactly; otherwise this returns
-	/// [`KeyParsingError::BadKeypair`]. This prevents importing a keypair whose
-	/// public key does not correspond to its secret key — which would otherwise
-	/// let an object sign with one key while advertising an unrelated public key
-	/// (e.g. a receive address the victim cannot spend from).
-	///
-	/// The secret key's internal invariants are checked as well: the stored
-	/// `t0` must match the low bits re-derived from `(rho, s1, s2)`, and the
-	/// stored `tr` must equal `SHAKE256(pk)`. Signing uses both fields, so a
-	/// blob corrupted in those regions would otherwise import cleanly and then
-	/// produce signatures that fail under the advertised public key.
-	///
-	/// Note: the `secret` and `public` fields are public, so callers can still
-	/// construct or mutate a `Keypair` with mismatched halves directly. This
-	/// check only guards the deserialization/import path.
-	pub fn from_bytes(bytes: &[u8]) -> Result<Keypair, KeyParsingError> {
-		if bytes.len() != SECRETKEYBYTES + PUBLICKEYBYTES {
-			return Err(KeyParsingError::BadKeypair);
-		}
-		let (secret_slice, public_bytes) = bytes.split_at(SECRETKEYBYTES);
-		// Copy the secret half into a zeroizing local (filled in place, so no
-		// intermediate `Copy` array is created). A plain `[u8; N]` local would
-		// survive the "move" into the returned struct — arrays are `Copy` —
-		// and leave a plaintext secret key in dead stack memory; `Zeroizing`
-		// wipes the local on every exit path, including the error returns.
-		let mut secret_bytes = Zeroizing::new([0u8; SECRETKEYBYTES]);
-		secret_bytes.copy_from_slice(secret_slice);
-		let public =
-			PublicKey::from_bytes(public_bytes).map_err(|_| KeyParsingError::BadKeypair)?;
-
-		// Enforce the cross-field invariants: the secret key must be internally
-		// consistent (tr, t0) and the public key must be the one that corresponds
-		// to it. The derived pk is public data, so a non-constant-time comparison
-		// is fine. The SecretKey is constructed directly (not via
-		// `SecretKey::from_bytes`) so the keygen-scale derivation runs once, not
-		// twice.
-		let derived_public = crate::sign::public_key_from_secret(&secret_bytes)
-			.ok_or(KeyParsingError::BadKeypair)?;
-		if derived_public != public.bytes {
-			return Err(KeyParsingError::BadKeypair);
-		}
-
-		Ok(Keypair { secret: SecretKey { bytes: *secret_bytes }, public })
-	}
-
-	/// Compute a signature for a given message.
-	///
-	/// # Arguments
-	///
-	/// * 'msg' - message to sign (max 64 MiB)
-	/// * 'ctx' - optional context string (max 255 bytes)
-	/// * 'hedge' - optional random bytes for hedged signing
-	///
-	/// # Errors
-	///
-	/// Returns `SignatureError::MessageTooLong` if the message exceeds 64 MiB.
-	/// Returns `SignatureError::ContextTooLong` if the context exceeds 255 bytes.
-	pub fn sign(
-		&self,
-		msg: &[u8],
-		ctx: Option<&[u8]>,
-		hedge: Option<[u8; params::SEEDBYTES]>,
-	) -> Result<Signature, SignatureError> {
-		self.secret.sign(msg, ctx, hedge)
-	}
-
-	/// Verify a signature for a given message with a public key.
-	///
-	/// # Arguments
-	///
-	/// * 'msg' - message that is claimed to be signed (max 64 MiB)
-	/// * 'sig' - signature to verify
-	/// * 'ctx' - optional context string (max 255 bytes)
-	///
-	/// Returns 'true' if the verification process was successful, 'false' otherwise.
-	/// Returns 'false' if the message exceeds 64 MiB, the context exceeds 255 bytes,
-	/// or the signature length is incorrect.
-	pub fn verify(&self, msg: &[u8], sig: &[u8], ctx: Option<&[u8]>) -> bool {
-		self.public.verify(msg, sig, ctx)
-	}
-}
-
-impl fmt::Debug for Keypair {
-	fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-		f.debug_struct("Keypair").field("public", &self.public).finish()
-	}
-}
-
-/// Private key.
-///
-/// `Clone` is intentionally not derived because the underlying bytes are sensitive.
-/// To explicitly copy a secret key, use `SecretKey::from_bytes(&sk.to_bytes())?`,
-/// which makes the duplication of secret material visible at every call site.
-///
-/// `Zeroize` is derived (in addition to `ZeroizeOnDrop`) so wrapper types that
-/// embed a `SecretKey` can themselves `#[derive(Zeroize, ZeroizeOnDrop)]`.
-#[derive(Zeroize, ZeroizeOnDrop)]
-pub struct SecretKey {
-	bytes: [u8; SECRETKEYBYTES],
-}
-
-impl SecretKey {
-	/// Returns a copy of the underlying bytes in a self-wiping buffer.
-	///
-	/// The copy is `Zeroizing`: it is erased when dropped, so callers no
-	/// longer need (or are able to forget) manual cleanup. Deref to
-	/// `[u8; SECRETKEYBYTES]` for access.
-	pub fn to_bytes(&self) -> Zeroizing<[u8; SECRETKEYBYTES]> {
-		Zeroizing::new(self.bytes)
-	}
-
-	/// Create a SecretKey from bytes.
-	///
-	/// # Arguments
-	///
-	/// * 'bytes' - private key bytes
-	///
-	/// Returns a SecretKey
-	///
-	/// # Consistency check
-	///
-	/// The packed key's internal invariants are validated, matching the
-	/// defense on the [`Keypair::from_bytes`] path: the stored `t0` must
-	/// equal the low bits re-derived from `(rho, s1, s2)`, and the stored
-	/// `tr` must equal `SHAKE256(pk)` for the re-derived public key. Signing
-	/// consumes both stored fields (`tr` is bound into the message digest,
-	/// `t0` into hint computation), so a corrupted or tampered blob would
-	/// otherwise import cleanly and then produce signatures that fail under
-	/// the corresponding public key. Rejecting it here fails fast at import
-	/// instead of creating a persistent signing outage — and ensures the same
-	/// serialized secret material has the same guarantees whether it is
-	/// imported as a `Keypair` or as a standalone `SecretKey`.
-	pub fn from_bytes(bytes: &[u8]) -> Result<SecretKey, KeyParsingError> {
-		if bytes.len() != SECRETKEYBYTES {
-			return Err(BadSecretKey);
-		}
-		// Zeroizing local filled in place: see `Keypair::from_bytes` — a bare
-		// `[u8; N]` local is `Copy` and would survive the move into the
-		// returned struct as an unwiped plaintext copy of the secret key.
-		let mut sk = Zeroizing::new([0u8; SECRETKEYBYTES]);
-		sk.copy_from_slice(bytes);
-		// Re-derives the public components and checks the stored tr/t0 against
-		// them; the derived pk itself is not needed here.
-		crate::sign::public_key_from_secret(&sk).ok_or(BadSecretKey)?;
-		Ok(SecretKey { bytes: *sk })
-	}
-
-	/// Compute a signature for a given message.
-	///
-	/// # Arguments
-	///
-	/// * 'msg' - message to sign (max 64 MiB)
-	/// * 'ctx' - context string (max 255 bytes)
-	/// * 'hedged' - wether to use RNG or not
-	///
-	/// # Errors
-	///
-	/// Returns `SignatureError::MessageTooLong` if the message exceeds 64 MiB.
-	/// Returns `SignatureError::ContextTooLong` if the context exceeds 255 bytes.
-	pub fn sign(
-		&self,
-		msg: &[u8],
-		ctx: Option<&[u8]>,
-		hedge: Option<[u8; params::SEEDBYTES]>,
-	) -> Result<Signature, SignatureError> {
-		if msg.len() > MAX_MESSAGE_SIZE {
-			return Err(SignatureError::MessageTooLong);
-		}
-		// The message is hashed as `domain_prefix || msg`. Only the small (<= 257
-		// byte) domain prefix is materialized; `msg` is passed by reference and
-		// absorbed directly, so an attacker-sized message is never copied.
-		match ctx {
-			Some(x) => {
-				if x.len() > 255 {
-					return Err(SignatureError::ContextTooLong);
-				}
-				let x_len = x.len();
-				let mut prefix = [0u8; 2 + 255];
-				prefix[1] = x_len as u8;
-				prefix[2..2 + x_len].copy_from_slice(x);
-				let mut sig: Signature = [0u8; SIGNBYTES];
-				crate::sign::signature(&mut sig, &prefix[..2 + x_len], msg, &self.bytes, hedge);
-				Ok(sig)
-			},
-			None => {
-				let mut sig: Signature = [0u8; SIGNBYTES];
-				// Prefix 2 zero bytes (domain_sep=0, context_len=0) for pure signatures
-				crate::sign::signature(&mut sig, &[0u8, 0u8], msg, &self.bytes, hedge);
-				Ok(sig)
-			},
-		}
-	}
-}
-
-#[derive(Eq, Clone, PartialEq, Debug, Hash, PartialOrd, Ord)]
-pub struct PublicKey {
-	pub bytes: [u8; PUBLICKEYBYTES],
-}
-
-impl PublicKey {
-	/// Returns a copy of underlying bytes.
-	pub fn to_bytes(&self) -> [u8; PUBLICKEYBYTES] {
-		self.bytes
-	}
-
-	/// Create a PublicKey from bytes.
-	///
-	/// # Arguments
-	///
-	/// * 'bytes' - public key bytes
-	///
-	/// Returns a PublicKey
-	pub fn from_bytes(bytes: &[u8]) -> Result<PublicKey, KeyParsingError> {
-		let bytes: [u8; PUBLICKEYBYTES] =
-			bytes.try_into().map_err(|_| KeyParsingError::BadPublicKey)?;
-
-		// Reject the degenerate all-zero t1 public key. With t1 = 0 the challenge term in the
-		// verification relation vanishes, letting an attacker forge signatures without any
-		// secret key (see `sign::verify`). Honest key generation never produces t1 = 0, so
-		// this only rejects malformed/malicious keys and never a legitimate one.
-		let mut rho = [0u8; params::SEEDBYTES];
-		let mut t1 = crate::polyvec::Polyveck::default();
-		crate::packing::unpack_pk(&mut rho, &mut t1, &bytes);
-		if t1.vec.iter().all(|p| p.coeffs.iter().all(|&c| c == 0)) {
-			return Err(KeyParsingError::BadPublicKey);
-		}
-
-		Ok(PublicKey { bytes })
-	}
-
-	/// Verify a signature for a given message with a public key.
-	///
-	/// # Arguments
-	///
-	/// * 'msg' - message that is claimed to be signed (max 64 MiB)
-	/// * 'sig' - signature to verify
-	/// * 'ctx' - context string (max 255 bytes)
-	///
-	/// Returns 'true' if the verification process was successful, 'false' otherwise.
-	/// Returns 'false' early if the message exceeds 64 MiB or context exceeds 255 bytes.
-	pub fn verify(&self, msg: &[u8], sig: &[u8], ctx: Option<&[u8]>) -> bool {
-		// Validate signature length first
-		let sig: &[u8; SIGNBYTES] = match sig.try_into() {
-			Ok(s) => s,
-			Err(_) => return false,
-		};
-		if msg.len() > MAX_MESSAGE_SIZE {
-			return false;
-		}
-		// As in `sign`, only the small domain prefix is materialized; the message
-		// is absorbed by reference rather than copied into a full-size buffer.
-		match ctx {
-			Some(x) => {
-				if x.len() > 255 {
-					return false;
-				}
-				let x_len = x.len();
-				let mut prefix = [0u8; 2 + 255];
-				prefix[1] = x_len as u8;
-				prefix[2..2 + x_len].copy_from_slice(x);
-				crate::sign::verify(sig, &prefix[..2 + x_len], msg, &self.bytes)
-			},
-			None => crate::sign::verify(sig, &[0u8, 0u8], msg, &self.bytes),
-		}
-	}
-}
+crate::frontend::define_ml_dsa!(crate::params::ml_dsa_87);
 
 #[cfg(test)]
 mod tests {
